Generalized the conditional rbm.
Rather than assuming the conditioned patterns will be the visible
units at previous time steps, we just take a size to condition on and
provide a constructor that takes the number of steps as a feature.

With the above change the split_vis method has changed and all references
to `history` or `hist` have been changed to `cond`.

Also, fixed an issue on line 51 which was passing `Bernoulli` instead of V and
H parametrics to the more general constructor.